+++ src/main/java/bridge/constants/Error.java
@@ -1,4 +1,4 @@
-package bridge.service.constants;
+package bridge.constants;
 
 public class Error {
     public static final String ERROR_MESSAGE = "[ERROR] ";
