+++ src/main/java/bridge/domain/Bridge.java
@@ -1,4 +1,4 @@
-package bridge.model;
+package bridge.domain;
 
 import bridge.domain.BridgeMaker;
