+++ src/main/java/bridge/domain/BridgeShape.java
@@ -1,4 +1,4 @@
-package bridge.model;
+package bridge.domain;
 
 import java.util.ArrayList;
 import java.util.List;
