+++ src/main/java/bridge/domain/GameControll.java
@@ -1,4 +1,4 @@
-package bridge.model;
+package bridge.domain;
 
 public class GameControll {
