+++ src/main/java/bridge/view/OutputView.java
@@ -1,8 +1,6 @@
 package bridge.view;
 
-import bridge.service.constants.Error;
-
-import static bridge.service.constants.Error.ERROR_MESSAGE;
+import static bridge.constants.Error.ERROR_MESSAGE;
 
 /**
  * 사용자에게 게임 진행 상황과 결과를 출력하는 역할을 한다.
