+++ src/test/java/bridge/model/BridgeShapeTest.java
@@ -1,11 +1,11 @@
 package bridge.model;
 
+import bridge.domain.BridgeShape;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
-import static org.junit.jupiter.api.Assertions.*;
 
 class BridgeShapeTest {
