+++ src/test/java/bridge/model/GameControllTest.java
@@ -1,5 +1,6 @@
 package bridge.model;
 
+import bridge.domain.GameControll;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
