+++ src/test/java/bridge/service/ValificationTest.java
@@ -1,5 +1,6 @@
 package bridge.service;
 
+import bridge.view.Valification;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
